:hammer: Bonjour protocol registration data to json serialization
