+++ composeApp/src/desktopMain/kotlin/com/clipevery/routing/SyncRouting.kt
@@ -5,7 +5,6 @@ import com.clipevery.dao.SyncInfoDao
 import com.clipevery.dto.sync.RequestSyncInfo
 import com.clipevery.dto.sync.SyncInfo
 import com.clipevery.signal.ClipIdentityKeyStore
-import com.clipevery.utils.decodePreKeyBundle
 import com.clipevery.utils.successResponse
 import io.ktor.server.application.call
 import io.ktor.server.request.receive
@@ -34,10 +33,9 @@ fun Routing.syncRouting() {
             for (requestSyncInfo in requestSyncInfos) {
                 val signalProtocolAddress = SignalProtocolAddress(requestSyncInfo.appInfo.appInstanceId, 1)
                 val sessionBuilder = SessionBuilder(sessionStore, preKeyStore, signedPreKeyStore, identityKeyStore, signalProtocolAddress)
-                val preKeyBundle = decodePreKeyBundle(requestSyncInfo.preKeyBundle)
+                val preKeyBundle = requestSyncInfo.preKeyBundle
                 syncInfoDao.saveSyncInfo(SyncInfo(requestSyncInfo.appInfo, requestSyncInfo.endpointInfo))
                 sessionBuilder.process(preKeyBundle)
-
             }
         }
         successResponse(call)
